Rename internal module and tests
