+++ tests/test_geometry/test_profiles/test_ub.py
@@ -16,7 +16,7 @@ from structuralcodes.geometry.profiles import (
 from structuralcodes.materials.basic import (
     ElasticMaterial,
 )
-from structuralcodes.sections._generic import BeamSection
+from structuralcodes.sections import BeamSection
 
 
 def load_ub_profiles_data():
